Рефакторинг работы с базой данных для пользователя
