+++ app/src/main/java/com/example/contacts/database/DataBaseComands.java
@@ -1,0 +1,7 @@
+package com.example.contacts.database;
+
+public enum DataBaseComands {
+    USER_ENTER,
+    USER_REGISTRY,
+    USER_CHECK_REGISTRY
+}
